agents/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
autogen-agentchat==0.4.2
autogen-core==0.4.2
autogen-ext==0.4.2
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
Deprecated==1.2.15
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.1.0
Flask-Cors==5.0.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jsonref==1.1.0
MarkupSafe==3.0.2
openai==1.59.8
opentelemetry-api==1.29.0
pillow==11.1.0
playwright==1.49.1
protobuf==4.25.5
pydantic==2.10.5
pydantic_core==2.27.2
pyee==12.0.0
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
tiktoken==0.8.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
